Move JobApplicationNotificationBody to the Notification components directory
