+++ src/components/Notification/JobApplicationNotificationBody/styles.module.scss
